stages interfaces and makes Dora the default abstract format
