+++ packages/neon-api/src/provider/dora/interface.ts
@@ -1,0 +1,17 @@
+export interface AddressAbstracts {
+  totalPages: number;
+  totalEntries: number;
+  pageSize: number;
+  pageNumber: number;
+  entries: Entry[];
+}
+
+export interface Entry {
+  txid: string;
+  time: number;
+  blockHeight: number;
+  asset: string;
+  amount: number;
+  addressTo: string;
+  addressFrom: string;
+}
