+++ packages/neon-api/src/provider/neonDB/class.ts
@@ -1,5 +1,5 @@
 import { logging, rpc, settings, u, wallet } from "@cityofzion/neon-core";
-import { AddressAbstract, PastTransaction, Provider } from "../common";
+import { PastTransaction, Provider } from "../common";
 import {
   getBalance,
   getClaims,
@@ -18,6 +18,7 @@ export class NeonDB implements Provider {
   }
 
   private rpc: rpc.RPCClient | null = null;
+
   private cacheExpiry: Date | null = null;
 
   public constructor(url: string) {
@@ -50,15 +51,19 @@ export class NeonDB implements Provider {
   public getBalance(address: string): Promise<wallet.Balance> {
     return getBalance(this.url, address);
   }
+
   public getClaims(address: string): Promise<wallet.Claims> {
     return getClaims(this.url, address);
   }
+
   public getMaxClaimAmount(address: string): Promise<u.Fixed8> {
     return getMaxClaimAmount(this.url, address);
   }
+
   public getHeight(): Promise<number> {
     return getHeight(this.url);
   }
+
   public getTransactionHistory(address: string): Promise<PastTransaction[]> {
     return getTransactionHistory(this.url, address);
   }
