+++ packages/neon-api/src/provider/neoscan/responses.ts
@@ -1,34 +1,18 @@
-import { AddressAbstract } from "../common";
+import {
+  AddressAbstractsString,
+  NeoscanBalance,
+  NeoscanClaim
+} from "./interface";
+
 export interface NeoscanV1GetBalanceResponse {
   balance: NeoscanBalance[] | null;
   address: string;
-}
-
-export interface NeoscanBalance {
-  asset: string;
-  amount: number;
-  unspent: NeoscanTx[];
-}
-
-export interface NeoscanTx {
-  txid: string;
-  value: number;
-  n: number;
 }
 
 export interface NeoscanV1GetClaimableResponse {
   unclaimed: number;
   claimable: NeoscanClaim[] | null;
   address: string;
-}
-
-export interface NeoscanClaim {
-  txid: string;
-  n: number;
-  value: number;
-  unclaimed: number;
-  start_height: number;
-  end_height: number;
 }
 
 export interface NeoscanV1GetUnclaimedResponse {
@@ -40,54 +24,4 @@ export interface NeoscanV1GetHeightResponse {
   height: number;
 }
 
-export interface NeoscanPastTx {
-  vouts: [
-    {
-      value: number;
-      transaction_id: number;
-      asset: string;
-      address_hash: string;
-    }
-  ];
-  vin: [
-    {
-      value: number;
-      txid: string;
-      n: number;
-      asset: string;
-      address_hash: string;
-    }
-  ];
-  type: string;
-  txid: string;
-  transfers: [
-    {
-      txid: string;
-      time: number;
-      contract: string;
-      block_height: number;
-      amount: number;
-      address_to: string;
-      address_from: string;
-    }
-  ];
-  time: number;
-  sys_fee: string;
-  size: number;
-  net_fee: string;
-  id: number;
-  claims: [
-    {
-      value: number;
-      txid: string;
-      n: number;
-      asset: string;
-      address_hash: string;
-    }
-  ];
-  block_height: number;
-  block_hash: string;
-  asset: null;
-}
-
-export type NeoscanAddressAbstracts = AddressAbstract;
+export type NeoscanGetAddressAbstractsResponse = AddressAbstractsString;
